Sprint 4 completion

+++ application/linkkivinkit/src/main/java/linkkivinkki/dao/InternetContentDao.java
@@ -8,10 +8,10 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
-import java.util.Map;
-import java.util.Map.Entry;
+import java.util.List;
+import linkkivinkki.domain.User;
 
-public class InternetContentDao implements Dao {
+public class InternetContentDao implements ItemDao {
 
     private Database database;
 
@@ -37,8 +37,9 @@ public class InternetContentDao implements Dao {
                 InternetContent content = new InternetContent(results.getString("title"), results.getString("url"), results.getString("description"));
                 content.setId(results.getInt("id"));
                 content.setCreationDate(results.getDate("date_created"));
-                // Other content related to the Item parent class should be inserted here when applicable
+                content.setUserId(results.getInt("user_id"));
 
+                // Other content related to the Item parent class should be inserted here when applicable
                 list.add(content);
             }
 
@@ -73,12 +74,13 @@ public class InternetContentDao implements Dao {
             Connection conn = database.getConnection();
 
             PreparedStatement addContent = conn.prepareStatement("INSERT INTO InternetContent "
-                    + "(title, url, description, date_created) "
-                    + "VALUES (?, ?, ?, ?);");
+                    + "(title, url, description, date_created, user_id) "
+                    + "VALUES (?, ?, ?, ?, ?);");
             addContent.setString(1, content.getTitle());
             addContent.setString(2, content.getUrl());
             addContent.setString(3, content.getDescription());
             addContent.setDate(4, new java.sql.Date(content.getCreationDate().getTime()));
+            addContent.setInt(5, i.getUserId());
             addContent.execute();
 
             // Close the connection
@@ -96,14 +98,16 @@ public class InternetContentDao implements Dao {
      * Deletes a InternetContent with the given id
      *
      * @param id - The id of the InternetContent to be deleted
+     * @param userId
      * @return True if successful, false if something went wrong
      */
     @Override
-    public boolean delete(int id) {
+    public boolean delete(int id, int userId) {
         try {
             Connection conn = database.getConnection();
-            PreparedStatement deleteContent = conn.prepareStatement("DELETE FROM InternetContent WHERE id = ?;");
+            PreparedStatement deleteContent = conn.prepareStatement("DELETE FROM InternetContent WHERE id = ? AND user_id = ?;");
             deleteContent.setInt(1, id);
+            deleteContent.setInt(2, id);
             deleteContent.execute();
 
             // Close the connection
@@ -136,6 +140,7 @@ public class InternetContentDao implements Dao {
                 found = new InternetContent(results.getString("title"), results.getString("url"), results.getString("description"));
                 found.setId(id);
                 found.setCreationDate(results.getDate("date_created"));
+                found.setUserId(results.getInt("user_id"));
             }
 
             results.close();
@@ -183,5 +188,37 @@ public class InternetContentDao implements Dao {
         }
 
         return true;
+    }
+
+    @Override
+    public List<InternetContent> findAllByUserId(int id) {
+        ArrayList<InternetContent> list = new ArrayList<>();
+
+        try {
+            Connection conn = database.getConnection();
+            PreparedStatement fetch = conn.prepareStatement("SELECT * FROM InternetContent where user_id = ?;");
+            fetch.setInt(1, id);
+            ResultSet results = fetch.executeQuery();
+
+            while (results.next()) {
+                InternetContent content = new InternetContent(results.getString("title"), results.getString("url"), results.getString("description"));
+                content.setId(results.getInt("id"));
+                content.setCreationDate(results.getDate("date_created"));
+                content.setUserId(id);
+                // Other content related to the Item parent class should be inserted here when applicable
+
+                list.add(content);
+            }
+
+            // Close the connection
+            results.close();
+            fetch.close();
+            conn.close();
+
+        } catch (SQLException ex) {
+            // Nothing happens, an empty list is returned at the end of the method
+        }
+
+        return list;
     }
 }
